backend/aws/lambda/requirements.txt
requests==2.23.0
psycopg2-binary==2.8.5